backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.11.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
PyJWT[crypto]==2.10.*
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.4
boto3==1.36.*
